+++ CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] / 2024-03-04
+- Update references
+- Test in Revit 2024
+
 ## [1.0.1] / 2022-04-26
 - Add Installation Files
 
 ## [1.0.0] / 2022-04-25
 - First Release
 
+[1.0.2]: ../../compare/1.0.1...1.0.2
 [1.0.1]: ../../compare/1.0.0...1.0.1
 [1.0.0]: ../../compare/1.0.0
