+++ RevitAddin.TemporaryGraphicsExample/Revit/App.cs
@@ -5,14 +5,14 @@ using System;
 
 namespace RevitAddin.TemporaryGraphicsExample.Revit
 {
-    [Console]
+    [AppLoader]
     public class App : IExternalApplication
     {
         private static RibbonPanel ribbonPanel;
         public Result OnStartup(UIControlledApplication application)
         {
             ribbonPanel = application.CreatePanel("Graphics");
-            ribbonPanel.AddPushButton<Commands.Command>("Temporary")
+            ribbonPanel.CreatePushButton<Commands.Command>("Temporary")
                 .SetLargeImage("/UIFrameworkRes;component/ribbon/images/revit.ico");
 
             return Result.Succeeded;
