+++ RevitAddin.TemporaryGraphicsExample/Revit/Commands/Command.cs
@@ -1,10 +1,8 @@
 using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
-using Autodesk.Revit.DB.ExternalService;
 using Autodesk.Revit.UI;
 using Autodesk.Revit.UI.Selection;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 
@@ -13,6 +11,7 @@ namespace RevitAddin.TemporaryGraphicsExample.Revit.Commands
     [Transaction(TransactionMode.Manual)]
     public class Command : IExternalCommand
     {
+        private Random Random = new Random();
         public Result Execute(ExternalCommandData commandData, ref string message, ElementSet elementSet)
         {
             UIApplication uiapp = commandData.Application;
@@ -27,7 +26,19 @@ namespace RevitAddin.TemporaryGraphicsExample.Revit.Commands
             {
                 temporaryGraphicsManager.Clear();
                 var point = selection.PickPoint();
-                var imagePath = Path.Combine(Location, "image.bmp");
+
+                var imageName = "image.bmp";
+                var random = Random.NextDouble();
+                if (random > 0.3)
+                {
+                    imageName = "image16.bmp";
+                }
+                if (random > 0.8)
+                {
+                    imageName = "device_power_browser.bmp";
+                }
+
+                var imagePath = Path.Combine(Location, imageName);
                 var data = new InCanvasControlData(imagePath, point);
                 var indexClick = temporaryGraphicsManager.AddControl(data, ElementId.InvalidElementId);
             }
@@ -36,50 +47,5 @@ namespace RevitAddin.TemporaryGraphicsExample.Revit.Commands
         }
 
         public string Location => Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-    }
-
-    public class ClickHandler : ITemporaryGraphicsHandler
-    {
-        public void AddServer()
-        {
-            Guid guid = GetServerId();
-            var services = (MultiServerService)ExternalServiceRegistry.GetService(GetServiceId());
-
-            if (services.IsRegisteredServerId(guid))
-                services.RemoveServer(guid);
-
-            services.AddServer(this);
-            services.SetActiveServers(new List<Guid> { guid });
-        }
-
-        public string GetDescription()
-        {
-            return "Click Temporary Graphics Handler";
-        }
-
-        public string GetName()
-        {
-            return "Click Temporary Graphics Handler";
-        }
-
-        public Guid GetServerId()
-        {
-            return new Guid("7B469077-9F7C-4CCF-9746-BD0DE41D3610");
-        }
-
-        public ExternalServiceId GetServiceId()
-        {
-            return ExternalServices.BuiltInExternalServices.TemporaryGraphicsHandlerService;
-        }
-
-        public string GetVendorId()
-        {
-            return "ricaun";
-        }
-
-        public void OnClick(TemporaryGraphicsCommandData data)
-        {
-            System.Windows.MessageBox.Show($"TemporaryGraphics {data.Index}");
-        }
     }
 }
